Getting upload and copy paste function part of draw.js
